% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{blink_summary}
\alias{blink_summary}
\title{Check blinks}
\usage{
blink_summary(data, Summary = "Event", ReturnData = FALSE)
}
\arguments{
\item{data}{A data table object output by \code{\link{create_time_series}}.}

\item{Summary}{A character string indicating the type of summary.}

\item{ReturnData}{A logical indicating whether to return a data table
containing Start Time information for each event.}
}
\value{
Summary information
}
\description{
\code{blink_summary} summarizes Eyelink marked blinks by Event, Subject, or Item.
}
\examples{
# Load example data
data("Pupilex3")

blink_summary(Pupilex3, Summary = "Event")

# Please see the vignettes for detailed example usage.
# vignette("PupilPre_Basic_Preprocessing", package="PupilPre")

}
