% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{butter_filter_app}
\alias{butter_filter_app}
\title{Plots the effect of Butterworth filtering by event.}
\usage{
butter_filter_app(data)
}
\arguments{
\item{data}{A data table object.}
}
\description{
\code{butter_filter_app} produces a plot of Butterworth filtered pupil data over the
original data to visually inspect the effect of different filter settings.
}
\examples{

if (interactive()) {

# Load example data
data("Pupilex5")

butter_filter_app(Pupilex5)
}

# Please see the vignettes for detailed example usage.
# vignette("PupilPre_Interpolation_and_Filtering", package="PupilPre")

}
