% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{plot_compare_app}
\alias{plot_compare_app}
\title{Plots comparison of Pupil and Pupil_Previous by event.}
\usage{
plot_compare_app(data)
}
\arguments{
\item{data}{A data table object.}
}
\description{
\code{plot_compare_app} produces a comparison plot of Pupil and
Pupil_Previous by event to visual changes.
}
\examples{
if (interactive()) {

# Load example data
data("Pupilex4")

plot_compare_app(Pupilex4)
}

# Please see the vignettes for detailed example usage.
# vignette("PupilPre_Basic_Preprocessing", package="PupilPre")

}
