% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive.R
\name{plot_summary_app}
\alias{plot_summary_app}
\title{Plots summary of subject or item.}
\usage{
plot_summary_app(data)
}
\arguments{
\item{data}{A data table object.}
}
\description{
\code{plot_summary_app} plots summary of a given subject or item.
}
\examples{
if (interactive()) {

# Load example data
data("Pupilex4")

plot_summary_app(Pupilex4)
}

# Please see the vignettes for detailed example usage.
# vignette("PupilPre_Basic_Preprocessing", package="PupilPre")

}
