% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ppl_check_eye_recording}
\alias{ppl_check_eye_recording}
\title{Check which eyes were recorded during the experiment}
\usage{
ppl_check_eye_recording(data)
}
\arguments{
\item{data}{A data table object output by \code{\link{create_time_series}}.}
}
\value{
Text feedback and instruction.
}
\description{
\code{ppl_check_eye_recording} quickly checks which eyes contain gaze data
either using the EYE_TRACKED column (if available) or the Right and
Left interest area columns. It prints a summary and
suggests which setting to use for the \code{Recording} parameter in the
function \code{\link{ppl_select_recorded_eye}}.
}
\examples{
# Load example data
data("Pupilex2")

ppl_check_eye_recording(data = Pupilex2)

# Please see the vignettes for detailed example usage.
# vignette("PupilPre_Basic_Preprocessing", package="PupilPre")

}
