% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{Pupildat}
\alias{Pupildat}
\title{This is a sample pupil size dataset included in the package}
\description{
This is a sample pupil size dataset included in the package
}
\author{
Aki Kyröläinen
}
\keyword{data}
