% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{Pupilex5}
\alias{Pupilex5}
\title{This is an example dataset to illustrate certain functionality}
\description{
This is an example dataset to illustrate certain functionality
}
\author{
Aki Kyröläinen
}
\keyword{data}
